ipa: libcamera: add support for ov8858 sensor
Currently, libcamera does not have information for the ov8858 sensor
used in the PinePhone Pro, a phone designed to run Linux.

This commit adds metadata, especially that sensor gain is reported and
set in 1/16 discrete increments.

For more information, see "5.8 manual exposure compensation/ manual
gain compensation" in [0] and the driver in [1].

[0] http://www.ahdsensor.com/uploadfile/202008/55322e75316871.pdf
[1] https://github.com/megous/linux/blob/orange-pi-5.19/drivers/media/i2c/ov8858.c

Signed-off-by: Nicholas Roth <nicholas@rothemail.net>